Aggiunti tutti i load indicator necessari, piccolo bug nella barra di ricerca che non riesco a riprodurre consistentemente (e quindi a fixare)

+++ app/src/main/java/com/musashi/claymore/spike/spike/Models.kt
@@ -19,7 +19,7 @@ data class Slot(
         var producer: Producer?=null,
         var rating: String?=null,
         var tecnicals: String?=null,
-        var time: Long?=null,
+        var time: Double?=null,
         var tips: String?=null,
         var type: String?=null
 )
@@ -30,7 +30,7 @@ data class SlotCard(
         var name: String?=null,
         var producer: String?=null,
         var rating: String?=null,
-        var time: Long?=null,
+        var time: Double?=null,
         var type: String?=null
 )
